Separate platform components into groups and items
