+++ Buform/Platforms/Android/Groups/Text/TextFormGroupHeaderViewHolder.cs
