+++ Buform/Platforms/Android/Items/Button/ButtonFormItemComponent.cs
@@ -5,7 +5,7 @@ namespace Buform;
 [Preserve(AllMembers = true)]
 [FormComponent]
 // ReSharper disable once UnusedType.Global
-public sealed class ButtonFormComponent : IFormComponent
+public sealed class ButtonFormItemComponent : IFormComponent
 {
     public void Register()
     {
