+++ Buform/Platforms/Ios/Groups/Text/TextFormGroupComponent.cs
@@ -1,0 +1,13 @@
+namespace Buform;
+
+[Preserve(AllMembers = true)]
+[FormComponent]
+// ReSharper disable once UnusedType.Global
+public sealed class TextFormGroupComponent : IFormComponent
+{
+    public void Register()
+    {
+        FormPlatform.RegisterGroupHeaderClass<TextFormGroup, TextFormGroupHeader>();
+        FormPlatform.RegisterGroupFooterClass<TextFormGroup, TextFormGroupFooter>();
+    }
+}
