+++ Buform/Platforms/Ios/Items/DateTime/DateTimeFormItemComponent.cs
@@ -1,0 +1,12 @@
+namespace Buform;
+
+[Preserve(AllMembers = true)]
+[FormComponent]
+// ReSharper disable once UnusedType.Global
+public sealed class DateTimeFormItemComponent : IFormComponent
+{
+    public void Register()
+    {
+        FormPlatform.RegisterItemClass<DateTimeFormItem, DateTimeFormCell>();
+    }
+}
