+++ Buform/Platforms/Ios/Items/Segments/SegmentsFormItemComponent.cs
@@ -3,7 +3,7 @@ namespace Buform;
 [Preserve(AllMembers = true)]
 [FormComponent]
 // ReSharper disable once UnusedType.Global
-public sealed class SegmentsFormComponent : IFormComponent
+public sealed class SegmentsFormItemComponent : IFormComponent
 {
     public void Register()
     {
