+++ Buform/Platforms/Ios/Items/Slider/SliderFormItemComponent.cs
@@ -3,7 +3,7 @@ namespace Buform;
 [Preserve(AllMembers = true)]
 [FormComponent]
 // ReSharper disable once UnusedType.Global
-public sealed class SliderFormComponent : IFormComponent
+public sealed class SliderFormItemComponent : IFormComponent
 {
     public void Register()
     {
