+++ Buform/Platforms/Ios/Items/Switch/SwitchFormCell.cs
