+++ Buform/Platforms/Ios/Items/Switch/SwitchFormItemComponent.cs
@@ -3,10 +3,10 @@ namespace Buform;
 [Preserve(AllMembers = true)]
 [FormComponent]
 // ReSharper disable once UnusedType.Global
-public sealed class ButtonFormComponent : IFormComponent
+public sealed class SwitchFormItemComponent : IFormComponent
 {
     public void Register()
     {
-        FormPlatform.RegisterItemClass<ButtonFormItem, ButtonFormCell>();
+        FormPlatform.RegisterItemClass<SwitchFormItem, SwitchFormCell>();
     }
 }
