+++ Buform/Platforms/Ios/Items/Text/TextFormItemComponent.cs
@@ -3,13 +3,10 @@ namespace Buform;
 [Preserve(AllMembers = true)]
 [FormComponent]
 // ReSharper disable once UnusedType.Global
-public sealed class TextFormComponent : IFormComponent
+public sealed class TextFormItemComponent : IFormComponent
 {
     public void Register()
     {
-        FormPlatform.RegisterGroupHeaderClass<TextFormGroup, TextFormGroupHeader>();
-        FormPlatform.RegisterGroupFooterClass<TextFormGroup, TextFormGroupFooter>();
-
         FormPlatform.RegisterItemClass<IMultilineTextFormItem, MultilineTextFormCell>();
         FormPlatform.RegisterItemClass<ITextFormItem, TextFormCell>();
     }
